// front/src/app/models/vacune.module.ts
// Generated by https://quicktype.io

export interface VacuneResponse {
    content: Vacune[];
    pageable: Pageable;
    last: boolean;
    totalElements: number;
    totalPages: number;
    size: number;
    number: number;
    sort: Sort;
    first: boolean;
    numberOfElements: number;
    empty: boolean;
}

export interface Vacune {
    id: string;
    name: string;
    description: string;
}

export interface Pageable {
    pageNumber: number;
    pageSize: number;
    sort: Sort;
    offset: number;
    unpaged: boolean;
    paged: boolean;
}

export interface Sort {
    empty: boolean;
    sorted: boolean;
    unsorted: boolean;
}

export interface EditVacuneResponse {
    id: string;
    name: string;
    description: string;
}